Initial commit
Another BDO fishbot that is just the continuation of the groundwork laid out by CrayonCode, but seeing as the author no longer supports the project I've decided to fork and continue to update it for the time being. As always, it will remain 100% free and has additional features such as reserving relic slots, updated processing with coal support, multiple language support and more. As always please refer to the readme for more details.